requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.1
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
PFTL==1.0
pillow==11.0.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
PyVISA==1.14.1
PyVISA-py==0.7.2
PyYAML==6.0.2
six==1.16.0
typing_extensions==4.12.2
